Move Validation component one folder up
