+++ src/components/Authentication/SignIn.js
@@ -6,7 +6,7 @@ import {compose} from 'recompose';
 
 import Button from '../Button/Button';
 import {PasswordForgetLink} from './PasswordForget';
-import {Validation, Validator, ValidationHelper} from './Validation/Validation';
+import {Validation, Validator, ValidationHelper} from './Validation';
 import {formReducer, errorReducer} from '../../helpers/validationHelper';
 
 import * as ROUTES from '../../constants/routes';
